move containers to common components
