+++ frontend/src/metabase/common/components/AdHocQuestionLoader.unit.spec.js
@@ -1,8 +1,9 @@
 import { render } from "__support__/ui";
 import { delay } from "__support__/utils";
-import { AdHocQuestionLoader } from "metabase/containers/AdHocQuestionLoader";
 import Question from "metabase-lib/v1/Question";
 import * as ML_Urls from "metabase-lib/v1/urls";
+
+import { AdHocQuestionLoader } from "./AdHocQuestionLoader";
 
 describe("AdHocQuestionLoader", () => {
   let loadQuestionSpy, loadMetadataSpy, mockChild;
