+++ frontend/src/metabase/common/components/CollectionName.jsx
