+++ frontend/src/metabase/common/components/NotFoundFallbackPage.tsx
@@ -2,10 +2,9 @@ import type { LocationDescriptor } from "history";
 import { replace } from "react-router-redux";
 import { useMount } from "react-use";
 
+import { NotFound } from "metabase/components/ErrorPages";
 import { connect } from "metabase/lib/redux";
 import { refreshCurrentUser } from "metabase/redux/user";
-
-import { NotFound } from "../components/ErrorPages";
 
 type DispatchProps = {
   refreshCurrentUser: () => any;
