+++ frontend/src/metabase/common/components/QuestionResultLoader.jsx
