+++ frontend/src/metabase/common/components/QuestionResultLoader.unit.spec.js
@@ -1,8 +1,9 @@
 import { setupCardQueryEndpoints } from "__support__/server-mocks";
 import { render } from "__support__/ui";
-import { QuestionResultLoader } from "metabase/containers/QuestionResultLoader";
 import Question from "metabase-lib/v1/Question";
 import { createMockDataset } from "metabase-types/api/mocks";
+
+import { QuestionResultLoader } from "./QuestionResultLoader";
 
 describe("QuestionResultLoader", () => {
   it("should load a result given a question", () => {
