+++ frontend/src/metabase/common/components/SavedQuestionLoader.unit.spec.js
@@ -6,7 +6,6 @@ import {
   setupUnauthorizedSchemaEndpoints,
 } from "__support__/server-mocks";
 import { renderWithProviders, screen } from "__support__/ui";
-import SavedQuestionLoader from "metabase/containers/SavedQuestionLoader";
 import Question from "metabase-lib/v1/Question";
 import {
   createMockCard,
@@ -14,6 +13,8 @@ import {
   createMockColumn,
 } from "metabase-types/api/mocks";
 import { createSampleDatabase } from "metabase-types/api/mocks/presets";
+
+import SavedQuestionLoader from "./SavedQuestionLoader";
 
 const databaseMock = createSampleDatabase();
