+++ frontend/src/metabase/common/components/SnippetCollectionName.tsx
