+++ frontend/src/metabase/common/components/dnd/DropArea.jsx
